DestroysTeamMemberships action implemented
