+++ app/services/cancels_course_membership.rb
@@ -6,6 +6,7 @@ require_relative "cancels_course_membership/destroys_grades"
 require_relative "cancels_course_membership/destroys_group_memberships"
 require_relative "cancels_course_membership/destroys_membership"
 require_relative "cancels_course_membership/destroys_submissions"
+require_relative "cancels_course_membership/destroys_team_memberships"
 
 module Services
   class CancelsCourseMembership
@@ -56,13 +57,6 @@ class CancelsCourseMembership
   def self.removes_flagged_users(membership)
     FlaggedUser.for_course(membership.course)
       .for_flagged(membership.user)
-      .destroy_all
-    self
-  end
-
-  def self.removes_team_memberships(membership)
-    TeamMembership.for_course(membership.course)
-      .for_student(membership.user)
       .destroy_all
     self
   end
